% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ascii_output.R
\name{ascii_data}
\alias{ascii_data}
\title{Specifies path to one of the data files in the package}
\usage{
ascii_data(file)
}
\arguments{
\item{file}{Name of file}
}
\value{
Path to file
}
\description{
Specifies path to one of the data files in the package
}
\examples{
print("hi")
}
