% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ascii_output.R
\name{ascii_grid}
\alias{ascii_grid}
\title{Converts an ASCII map to a matrix}
\usage{
ascii_grid(image_map)
}
\arguments{
\item{image_map}{Map from ascii_map}
}
\value{
A matrix
}
\description{
Converts an ASCII map to a matrix
}
\examples{
print("hi")
}
