% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ascii_map.R
\name{ascii_map}
\alias{ascii_map}
\title{Creates ASCII art from an image}
\usage{
ascii_map(filename, alphabet = letters, rescale = NULL,
  threshold = 0.5)
}
\arguments{
\item{filename}{A character string specifying the path to the file}

\item{alphabet}{A character string that lists the set of characters to use}

\item{rescale}{Scale to resize image to (if NULL, sets maximum size of 100x100)}

\item{threshold}{Lightness value at which to truncate}
}
\value{
A data frame with...
}
\description{
Creates ASCII art from an image
}
\examples{
print("hi")
}
