% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ascii_output.R
\name{ascii_plot}
\alias{ascii_plot}
\title{Plots an ASCII map}
\usage{
ascii_plot(image_map, charsize = 4)
}
\arguments{
\item{image_map}{Map from ascii_map}

\item{charsize}{Size of the characters}
}
\value{
A ggplot object
}
\description{
Plots an ASCII map
}
\examples{
print("hi")
}
