% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ascii_output.R
\name{ascii_rain}
\alias{ascii_rain}
\title{Writes an ASCII grid to an HTML file with the rain animation}
\usage{
ascii_rain(text_grid, file)
}
\arguments{
\item{text_grid}{Matrix from ascii_grid}

\item{file}{Path to HMTL file}
}
\value{
A matrix, invisibly
}
\description{
Writes an ASCII grid to an HTML file with the rain animation
}
\examples{
print("hi")
}
