% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ascii_output.R
\name{ascii_text}
\alias{ascii_text}
\title{Writes an ASCII grid to text file}
\usage{
ascii_text(text_grid, file)
}
\arguments{
\item{text_grid}{Matrix from ascii_grid}

\item{file}{Path to text file}
}
\value{
A matrix, invisibly
}
\description{
Writes an ASCII grid to text file
}
\examples{
## Not run:
bayes_img <- ascii_data("bayes.png") # path to the bayes image
bayes_map <- ascii_map(bayes_img)    # construct ASCII map
bayes_grid <- ascii_grid(bayes_map)   # make grid
ascii_text(bayes_grid, file = "bayes_grid.txt")
## End(Not run)
}
