% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ascii_output.R
\name{ascii_data}
\alias{ascii_data}
\title{Specifies path to one of the data files in the package}
\usage{
ascii_data(file)
}
\arguments{
\item{file}{Name of file as a character string}
}
\value{
Path to file as a character string
}
\description{
Specifies path to one of the data files in the package
}
\details{
This is a convenience function that returns the path to one of the
external data files bundled in the asciify package. There are only four 
such files: "bayes.png", "unicorn.png", "heart.jpg", "rain.html"
}
\examples{
ascii_data("bayes.png")
ascii_data("unicorn.png")
ascii_data("heart.jpg")
ascii_data("rain.html")
}
