% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/asciify.R
\name{ascii_map}
\alias{ascii_map}
\title{Creates ASCII art from an image}
\usage{
ascii_map(filename, alphabet = c(LETTERS, letters), threshold = 0.5)
}
\arguments{
\item{filename}{A character string specifying the path to the file}

\item{alphabet}{A character string that lists the set of characters to use}

\item{threshold}{Lightness value at which to truncate}
}
\value{
A data frame with...
}
\description{
Creates ASCII art from an image
}
\examples{
print("hi")
}
