% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ascii_output.R
\name{ascii_rain}
\alias{ascii_rain}
\title{Writes an ASCII grid to an HTML file with the rain animation}
\usage{
ascii_rain(text_grid, file, fontsize = "5px", lineheight = "4px",
  turnon = 0.1, turnoff = 0.025)
}
\arguments{
\item{text_grid}{Matrix from ascii_grid}

\item{file}{Path to HMTL file}

\item{fontsize}{How big is the text}

\item{lineheight}{How tall is a line}

\item{turnon}{Animation parameter}

\item{turnoff}{Animation parameter}
}
\value{
A matrix, invisibly
}
\description{
Writes an ASCII grid to an HTML file with the rain animation
}
\examples{
## Not run:
bayes_img <- ascii_data("bayes.png") # path to the bayes image
bayes_map <- ascii_map(bayes_img)    # construct ASCII map
bayes_grid <- ascii_grid(bayes_map)   # make grid
ascii_rain(bayes_grid, file = "bayes_rain.html")
## End(Not run)
}
