% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ascii_output.R
\name{ascii_text}
\alias{ascii_text}
\title{Writes an ASCII grid to text file}
\usage{
ascii_text(text_grid, file)
}
\arguments{
\item{text_grid}{Matrix from ascii_grid}

\item{file}{Path to text file}
}
\value{
A matrix, invisibly
}
\description{
Writes an ASCII grid to text file
}
\examples{
print("hi")
}
